+++ client/src/App.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import logo from './logo.svg';
-import './App.css';
-
+import './css/App.css';
 class App extends Component {
   render() {
     return (
